refactoring message processing out of function endpoint
