+++ src/NServiceBus.AzureFunctions.InProcess.ServiceBus/Serverless/ServerlessInterceptor.cs
@@ -6,6 +6,6 @@
 
         public ServerlessInterceptor(ServerlessTransport transport) => this.transport = transport;
 
-        public PipelineInvoker PipelineInvoker => transport.PipelineInvoker;
+        public IMessageProcessor MessageProcessor => transport.MessageProcessor;
     }
 }
